[feat] 주문 금액 - 포인트 적용 구현

+++ src/main/java/com/c3l2/persome/order/dto/request/OrderRequestDto.java
@@ -27,7 +27,7 @@ public class OrderRequestDto {
     private String requestMessage;
 
     private Long userCouponId;
-    private BigDecimal usePointAmount;
+    private Integer usePointAmount;
 
     public Order toEntity(User user) {
         return Order.builder()
